# mobile/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anupalan API",
    "summary": "Compliance engine for packaged commodities in India",
    "description": "Advisory pre-audit tool. Output is not a certification and carries no legal force. The rule pack is an engineering transcription pending legal review.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/scans/{scan_id}/report": {
      "post": {
        "tags": ["reports"],
        "summary": "Generate a report for a scan",
        "description": "Render this scan's current findings and store them (FR-27, FR-08).\n\nRefuses a scan that has never been evaluated. There is nothing to report on, and a document\nstating no findings would read as a clean result rather than as an absent one.",
        "operationId": "create_report_v1_scans__scan_id__report_post",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reports/{report_id}": {
      "get": {
        "tags": ["reports"],
        "summary": "Fetch a generated report",
        "description": "One report, with fresh read URLs.\n\nThe links are minted per request and expire, so this is also how a client re-shares a report it\ngenerated yesterday \u2014 nothing hands out a durable link to evidence.",
        "operationId": "get_report_v1_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportCreateIn": {
        "properties": {
          "formats": {
            "items": {
              "type": "string",
              "enum": ["pdf", "docx", "json"]
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Formats",
            "description": "At least one. Duplicates are collapsed; order does not matter."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ReportCreateIn",
        "description": "Which documents to render."
      },
      "ReportFileOut": {
        "properties": {
          "format": {
            "type": "string",
            "enum": ["pdf", "docx", "json"],
            "title": "Format"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Time-limited read URL. Buckets are private (architecture \u00a710), so this is the only read path and it expires \u2014 fetch the bytes, do not store the link."
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          }
        },
        "type": "object",
        "required": ["format", "url", "size_bytes", "media_type"],
        "title": "ReportFileOut",
        "description": "One rendered document."
      },
      "ReportOut": {
        "properties": {
          "report_id": {
            "type": "string",
            "format": "uuid",
            "title": "Report Id"
          },
          "scan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Scan Id"
          },
          "evaluation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evaluation Id",
            "description": "Which evaluation this report states. A later correction makes a new revision and does not change what this document said."
          },
          "status": {
            "type": "string",
            "enum": ["pending", "ready", "failed"],
            "title": "Status",
            "description": "``ready`` as soon as it exists. ``pending`` is reserved for the asynchronous form \u2014 see this module's docstring \u2014 so a polling client is correct either way."
          },
          "rulepack_version": {
            "type": "string",
            "title": "Rulepack Version"
          },
          "formats": {
            "items": {
              "type": "string",
              "enum": ["pdf", "docx", "json"]
            },
            "type": "array",
            "title": "Formats",
            "description": "What was asked for. Kept beside ``files`` on purpose: a report delivering one of two requested documents must be visible as a short delivery rather than reading as though only one was ever wanted."
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/ReportFileOut"
            },
            "type": "array",
            "title": "Files"
          },
          "image_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Sha256",
            "description": "SHA-256 of the raw image as received. Null when the scan has no raw asset \u2014 which the client shows as an unverifiable image rather than as a verified one."
          },
          "findings_sha256": {
            "type": "string",
            "title": "Findings Sha256"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "scan_id",
          "evaluation_id",
          "status",
          "rulepack_version",
          "formats",
          "findings_sha256",
          "requested_at"
        ],
        "title": "ReportOut",
        "description": "A generated report set."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}